+++ CHANGES.md
@@ -11,6 +11,7 @@ Fast and flexible PHP library for text tables.
   * `RendererContract`'s `render()` no longer expects returns `string[]`.
   * Writers and `WriterContract` are completely gone as they are no longer needed.
   * `Cell`'s constructor allows to set title/cell align separately.
+  * `Column`'s `show()` and `hide()` helpers are now removed.
 
 
 * 1.7.0 (2022-10-19)